Move Ecto.Type to own module as macro
Revert "Move Ecto.Type to own module as macro"

This reverts commit 203017473b47d9e268a76d2be4087e222c6534a3.

Move Ecto.Type to own module as macro
